Add dynamic & static loading feature
