+++ package/collection2/dynamic.js
@@ -1,0 +1,3 @@
+Collection2.load = function () {
+    import './main';
+}
