+++ package/collection2/static.js
@@ -1,0 +1,1 @@
+import './main';
