+++ tests/autoValue.tests.js
@@ -1,3 +1,4 @@
+import 'meteor/aldeed:collection2/dynamic';
 import { Meteor } from 'meteor/meteor';
 import expect from 'expect';
 import { Mongo } from 'meteor/mongo';
@@ -5,6 +6,8 @@ import SimpleSchema from "meteor/aldeed:simple-schema";
 import { callMongoMethod } from './helper';
 
 /* global describe, it */
+
+Collection2.load();
 
 const collection = new Mongo.Collection('autoValueTestCollection');
 const localCollection = new Mongo.Collection('autoValueTestLocalCollection', {
